Restructure mbrl.util package
